I've changed un UNBELIEVABLE amount of things
Since the last update, every cell now has a certain number assigned to it, cells can be flagged, and when a zero cell is pressed, it opens up all adjacent cells.
Furthermore, mines are only generated once the initial cell is clicked, this prevents situations in which the player instantly clicks on a mine, or a number, preventing the game from being played at it's best.